+++ src/main/java/com/emeraldElves/alcohollabelproject/DatabaseController.java
@@ -73,7 +73,7 @@ public class DatabaseController {
                     new Database.TableField("brandName", "VARCHAR (10000) NOT NULL"),
                     new Database.TableField("origin", "INTEGER NOT NULL"),
                     new Database.TableField("type", "INTEGER NOT NULL"),  //only beer or wine?
-                    new Database.TableField("pH", "INTEGER"),
+                    new Database.TableField("pH", "REAL"),
                     new Database.TableField("vintageYear", "VARCHAR (255)"));
             Log.console("Created new AlcoholInfo table");
         } catch (SQLException e) {
